Add internationalization
The bulk of the changes in this commit were made by simply following the
instructions in the next-intl 'App Router setup without I18n routing'
guide [1]. The only exception being the resolution of locale based on
the Accept-Language header which is not covered in the guide.

For that I used the same `Negotiator` and `@formatjs/intl-localematcher`
packages as the next-intl middleware uses for the same purpose [2].

I initially tried to use the `Negotiator.language()` function instead of
taking a the dependency on `@formatjs/intl-localematcher`. However, when
using `Negotiator.language()`, I hit upon some very counterintuitive
behaviour (such as "en-GB,fr-CH;q=0.7,en;q=0.3" resolving to "fr") that
is currently considered 'as intended' by the package maintainers [3] and
therefore ended up using `@formatjs/intl-localematcher` as well.

Note also that the next-intl middleware applies some reordering of
available locales [4] as a workaround for a separate `@formatjs` issue
[5]. I have not done the same here as the two available locales have the
same specificity.

[1] https://next-intl.dev/docs/getting-started/app-router/without-i18n-routing
[2] https://github.com/amannn/next-intl/blob/79c20855d523cb9d467bac08969244928596ced0/packages/next-intl/src/middleware/resolveLocale.tsx#L40
[3] https://github.com/jshttp/negotiator/issues/54#issuecomment-473358325
[4] https://github.com/amannn/next-intl/blob/79c20855d523cb9d467bac08969244928596ced0/packages/next-intl/src/middleware/resolveLocale.tsx#L28
[5] https://github.com/formatjs/formatjs/issues/4469