Add internationalization
The existing ASP.NET MVC-based Buttercup application is available in
English and French. When signed in, the locale is determined by the
language setting on the user profile; otherwise, the locale is selected
based on the Accept-Language request header.

This commit uses next-intl [1] to add similar internationalization
support to the Next.js app, with the following notable differences:

1) In the Next.js app, the locale ID is always included in the route.
   Middleware redirect requests that don't include a locale ID to the
   equivalent route containing the most appropriate supported locale ID
   based on the Accept-Language request header.
2) Since we've yet to implement authentication in the Next.js app, we
   don't yet have any pages where the locale is determined by the user
   profile.

The changes in this commit were made by following the instructions in
the next-intl 'App Router setup with I18n routing' guide [2].

[1] https://next-intl.dev/
[2] https://next-intl.dev/docs/getting-started/app-router/with-i18n-routing