Add Sass module for breakpoints
Up to now I've been able to use vanilla CSS and custom CSS properties
to define common colours and dimensions.

However CSS custom properties cannot be used in media queries, so to
avoid having to hardcode the breakpoints wherever they appear, I've
decided to take a development dependency on Sass and define them as Sass
variables.

At this point we only have one breakpoint but I more to be added in due
course.