Add GraphQL schema
We'll use this schema to generate typed GraphQL operations.

It was downloaded from a local instance of the GraphQL server using the
`graphql:fetch-schema` npm script added in the previous commit.

In my opinion there's little benefit in reformatting files that are not
intended to be manually edited, so I've added the schema file to
Prettier's ignore list instead.