Run end-to-end tests in CI
The 'Install Playwright browsers' step is relatively time consuming,
currently taking longer than all the other steps in the pipeline put
together.

One way to avoid having this step is to run the tests in a containerized
job [1]. However, since the official Playwright docker image [2] doesn't
include all the dependencies (e.g. .NET and MySQL Server) that we'll
eventually need to run the end-to-end tests, this would involve
maintaining and deploying our own custom Docker image which is likely to
be more time-consuming and inconvenient overall.

[1] https://playwright.dev/docs/ci#docker
[2] https://github.com/microsoft/playwright/blob/main/utils/docker/Dockerfile.noble