Run end-to-end tests against production server in CI
In development, the speed and convenience of running the tests against
the development server (for example, not having to recompile the
application every time you modify the application before test runs)
generally outweigh the benefits of running the tests against a
production build.

In CI, however, the tradeoffs are reversed. There are no edits to watch,
correctness and consistency take priority over speed, and we have to run
`next build` anyway to check for compilation and typing errors.

This commit therefore updates the Playwright config to start the
production server when running the end-to-end tests in CI while
continuing to use the development server by default when running the
end-to-end tests locally.